start refactoring expensive action logic
