+++ app/Platform/Events/GameAwardMetricsUpdated.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Platform\Events;
+
+use App\Models\Game;
+use Illuminate\Broadcasting\InteractsWithSockets;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class GameAwardMetricsUpdated
+{
+    use Dispatchable;
+    use InteractsWithSockets;
+    use SerializesModels;
+
+    public function __construct(
+        public Game $game,
+    ) {
+    }
+
+    public function broadcastOn(): PrivateChannel
+    {
+        return new PrivateChannel('channel-name');
+    }
+}
