+++ app/Platform/Jobs/UpdateGameAwardMetricsJob.php
@@ -1,0 +1,54 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Platform\Jobs;
+
+use App\Models\Game;
+use App\Platform\Actions\UpdateGameAwardMetricsAction;
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldBeUniqueUntilProcessing;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
+
+class UpdateGameAwardMetricsJob implements ShouldQueue, ShouldBeUniqueUntilProcessing
+{
+    use Dispatchable;
+    use InteractsWithQueue;
+    use Queueable;
+    use SerializesModels;
+
+    public function __construct(
+        private readonly int $gameId,
+    ) {
+    }
+
+    public int $uniqueFor = 3600;
+
+    public function uniqueId(): string
+    {
+        return config('queue.default') === 'sync' ? '' : $this->gameId;
+    }
+
+    /**
+     * @return array<int, string>
+     */
+    public function tags(): array
+    {
+        return [
+            Game::class . ':' . $this->gameId,
+        ];
+    }
+
+    public function handle(): void
+    {
+        $game = Game::find($this->gameId);
+        if (!$game) {
+            return;
+        }
+
+        app()->make(UpdateGameAwardMetricsAction::class)->execute($game);
+    }
+}
